build: removed all build stuff from inside /web

+++ build.sh
@@ -3,4 +3,11 @@ if [ ! -f ./.env.example ]; then
   exit 1
 fi
 
+if [ ! -f ./web/.env.example ]; then
+  echo "\"./web/.env.example\" not found"
+  exit 1
+fi
+
+
 cp .env.example .env
+cp ./web/.env.example ./web/.env
